// src/app/@core/models/car.model.ts
import { Maintenance } from "./maintenance.model";

export interface Car {
  id?: number;
  plateNumber?: string;
  plateNumberIsVisible?: boolean;
  brand?: string;
  brandIsVisible?: boolean;
  model?: string;
  modelIsVisible?: boolean;
  variant?: string;
  variantIsVisible?: boolean;
  type?: string;
  typeIsVisible?: boolean;
  version?: string;
  versionIsVisible?: boolean;
  firstRegistration?: string;
  firstRegistrationIsVisible?: boolean;
  vinNumber?: string;
  vinNumberIsVisible?: boolean;
  fuel?: string;
  fuelIsVisible?: boolean;
  engineBrand?: string;
  engineBrandIsVisible?: boolean;
  engineCode?: string;
  engineCodeIsVisible?: boolean;
  cilinderAmount?: string;
  cilinderAmountIsVisible?: boolean;
  cilinderVolume?: string;
  cilinderVolumeIsVisible?: boolean;
  potencyKW?: string;
  potencyKWIsVisible?: boolean;
  potencyCV?: string;
  potencyCVIsVisible?: boolean;
  tires?: string;
  tiresIsVisible?: boolean;
  category?: string;
  categoryIsVisible?: boolean;
  emisionLevel?: string;
  emisionLevelIsVisible?: boolean;
  radioCode?: null;
  radioCodeIsVisible?: boolean;
  keyCode?: null;
  keyCodeIsVisible?: boolean;
  maintenanceReset?: string;
  maintenanceResetIsVisible?: boolean;
  observations?: string;
  observationsIsVisible?: boolean;
  created_at?: string;
  updated_at?: string;
  maintenances?: Maintenance[];
}